nested projects now work like folders
